Move pulp related code to `galaxy.worker`
* Move `galaxy.pulp.tasks` to `galaxy.worker.tasks.collection`
* Move `galaxy.pulp.exceptions` to `galaxy.worker.exceptions`
