+++ galaxy/worker/tasks/__init__.py
@@ -15,11 +15,11 @@
 # You should have received a copy of the Apache License
 # along with Galaxy.  If not, see <http://www.apache.org/licenses/>.
 
-
-class PulpTaskError(Exception):
-    """Base class for pulp task exceptions"""
-    pass
+from .collection import import_collection  # noqa: F401
+from .repository import import_repository  # noqa: F401
 
 
-class VersionConflict(PulpTaskError):
-    pass
+__all__ = (
+    'import_collection',
+    'import_repository',
+)
